Introduce tests for threadio.
This commit moves the synthetic data creation that was previously done in
the dataformat test suite into its own python class.  The tests there
are modified to use it.  Further, this data generation is reused to create
temporary files in the threadIO test suite.  Tests are added to validate
threadio functionality for the following threadio classes:
Image2D
Tensor2D
Tensor3D
SparseTensor2D
SparseTensor3D

Each is tested with 1 and 2 threads, 1 and 2 batch storage, making/not
making copies of the data in dataloader.

Every test is using the larcv interface, so that too is tested by this suite.

This closes issue #18.