Change data_generator to not re-import larcv.
Critical change: add a pause to allow the manager thread for larcv ThreadIO to start up when first calling start_manager in dataloader2.  If this is not present, and the system is oversubcribed, the start_manager function will return and a race condition will occur.  The manager is trying to start and fill data yet the dataloader continues ahead to read data.  This causes a crash and tests fail randomly and inexplicably.